api/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
keras==2.6.0
Keras-Preprocessing==1.1.2
matplotlib==3.4.3
numpy==1.19.5
Pillow==8.3.2
tensorflow==2.6.0
